ffi: cleanup pymem
